chore: cleanup index.ts file
